modfile: put new tests in replace_test.go
